+++ packages/solidart/example/main.dart
@@ -5,10 +5,10 @@ import 'dart:async';
 import 'package:solidart/solidart.dart';
 
 Future<void> main() async {
-  final count = createSignal(0);
-  final doubleCount = createComputed(() => count() * 2);
+  final count = Signal(0);
+  final doubleCount = Computed(() => count() * 2);
 
-  createEffect((dispose) {
+  Effect((dispose) {
     print('The counter is now: ${count()}');
     print('The double counter is now: ${doubleCount()}');
   });