+++ packages/solidart/CHANGELOG.md
@@ -22,6 +22,7 @@
 - **CHORE**: Rename the `firstWhere` method of a `ReadSignal` into `until`
 - **CHORE**: Rename the `firstWhereReady` method of a `Resource` into `untilReady`
 - **CHORE**: Rename the `update` method of a `Signal` into `updateValue`
+- **CHORE**: Deprecate `createSignal`, `createComputed`, `createEffect` and `createResource`
 
 ## 1.1.0